added unicode check

+++ WebGoatCore/Controllers/HomeController.cs
@@ -40,30 +40,50 @@ namespace WebGoatCore.Controllers
             ViewBag.Message = "";
             try
             {
-                string filename = Path.GetFileName(FormFile.FileName);
-                string webRootPath = _webHostEnvironment.WebRootPath;
-                string path = Path.Combine(webRootPath, "upload", filename);
+                // Create a temporary filename with .txt extension
+                string newFilename = $"{Path.GetRandomFileName()}{Guid.NewGuid()}.txt"; ;
+                string tempFolderPath = GetTemporaryDirectory();
+
+                // Generate a path with the filename
+                string path = Path.Combine(tempFolderPath, newFilename);
+
+                // Copy the contents of the file to the new location
                 using (var fileStream = new FileStream(path, FileMode.Create))
                 {
                     await FormFile.CopyToAsync(fileStream);
                 }
-                int firstNum = 0;
-                int secondNum = 0;
+
+                // Read the contents of the copied file
+                string? content;
                 using (StreamReader sr = new StreamReader(path))
                 {
-                    firstNum = int.Parse(sr.ReadLine());
-                    secondNum = int.Parse(sr.ReadLine());
+                    content = sr.ReadToEnd();
                 }
-                int sum = firstNum + secondNum;
-                System.IO.File.Delete(path);
-                ViewBag.Message = $"<div class='success' style='text-align:center'>The sum is {sum} </div>";
-                return View("About");
+                
+                // Check whether the content value is present or not
+                if (content != null && content.Length > 0)
+                {
+                    // Verify whether the content has unicode characters or not
+                    bool hasUnicode = System.Text.Encoding.UTF8.GetByteCount(content) != content.Length;
+
+                    // Throw error if the contents of the file contains unicode characters.
+                    if (hasUnicode == true)
+                    {
+                        throw new InvalidDataException("The given brochure contains some invalid characters. Please remove them.");
+                    }
+
+                    // Clean up resources
+                    System.IO.File.Delete(path);
+                    Directory.Delete(tempFolderPath, true);
+
+                    ViewBag.Message = "Successfully uploaded your feedback! Thank you!";
+                }
             }
             catch (Exception ex)
             {
-                ViewBag.Message = $"<div class='error' style='text-align:center'>File processing failed: {ex.Message} </div>";
-                return View("About");
+                ViewBag.Message = $"File processing failed: {ex.Message}";
             }
+            return View("About");
         }
 
         [Authorize(Roles = "Admin")]
@@ -77,6 +97,14 @@ namespace WebGoatCore.Controllers
                 RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier,
                 ExceptionInfo = HttpContext.Features.Get<IExceptionHandlerPathFeature>(),
             });
+        }
+
+        // Utility method to generate a temporary directory
+        private string GetTemporaryDirectory()
+        {
+            string tempDirectory = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
+            Directory.CreateDirectory(tempDirectory);
+            return tempDirectory;
         }
     }
 }
